replacing yield tokens sources with a proxy
